feat: add navigation menu component with viewport and indicator
feat: update select component with improved styling and structure

feat: create skeleton component for loading states

feat: enhance table component for responsive design with mobile cards

fix: update textarea component for consistent styling

feat: implement user menu for navigation with authentication state

fix: refactor user navigation component for better structure

fix: adjust database imports for consistency

fix: update schema to use numeric type for price fields

feat: add custom hook for fetching collections with error handling

fix: improve user engagement tracking with metrics

fix: streamline alerting and logging imports

fix: reorganize API middleware imports for clarity

feat: enhance atoms for better state management in the application

feat: add utility functions for formatting and error handling

fix: update types for dialog modal props to include additional attributes

+++ src/db/schema.ts
@@ -1,5 +1,5 @@
-import { pgTable, text, integer, serial, timestamp, varchar } from 'drizzle-orm/pg-core';
 import { relations, InferModel } from 'drizzle-orm';
+import { pgTable, text, integer, serial, timestamp, varchar, numeric } from 'drizzle-orm/pg-core';
 
 // Users table
 export const users = pgTable('users', {
@@ -18,7 +18,7 @@ export const collections = pgTable('collections', {
   id: serial('id').primaryKey().notNull(),
   name: text('name').notNull(),
   descriptions: text('descriptions'),
-  price: integer('price').notNull(),
+  price: numeric('price').notNull(),
   stocks: integer('stocks').notNull(),
   status: text('status', { enum: ['open', 'closed'] })
     .notNull()
@@ -33,7 +33,7 @@ export const collections = pgTable('collections', {
 // Bids table
 export const bids = pgTable('bids', {
   id: serial('id').primaryKey().notNull(),
-  price: integer('price').notNull(),
+  price: numeric('price').notNull(),
   collectionId: integer('collection_id')
     .notNull()
     .references(() => collections.id, { onDelete: 'cascade' }),
